SMB/BMB changes for calving front cells
I made several changes to handle the application of the surface and basal mass balances
in calving-front cells. Suppose a CF cell has thck = 20 m, with thck_calving_front = 200 m.
The effective fractional area coverage in the cell is 20/200 = 0.1. So we want to apply
the SMB and BMB to only 10% of the cell instead of the entire cell. The mass conservation
checks need to take this into account.

In glissade_tstep, CISM now computes effective_fracarea = thck/thck_calving_front
(within the range 0 to 1). This fraction is passed to the transport solver and
used for the SMB/BMB calculation. In particular, the fields acab_applied and bmlt_applied
(with track the actual mass balance applied to ice) are weighted by effective_areafrac.

I also made some logic changes so that acab_applied and bmlt_applied are accumulated correctly
when the transport calculation is subcycled.

There can still be mass conservation errors in a few cells as a result of effective_areafrac
changing values between when it is computed (in glissade_tstep) and when it is applied
(after horizontal transport). These errors will be fixed in a followup commit by applying
SMB and BMB before horizontal transport.

This commit is answer-changing for the subgrid calving front scheme (which_ho_calving_front = 1),
but not for default settings.