Moved SMB/BMB calculations before horizontal transport
Subroutine glissade_transport_driver has two main functions: to transport the
ice in the horizontal direction (followed by a vertical remapping of layers back
to sigma coordinates), and to add and remove ice at the upper and lower surfaces
due to nonzero acab or bmlt. Previously, horizontal transport was done before
applying the surface and basal mass balance.

With this commit, the order of calculations is reversed; the mass balance is applied
before horizontal transport. This allows us to compute effective_areafrac and the
ocean mask in glissade_tstep, pass then into the transport driver, and use them for the
SMB/BMB calculation, without having to recompute them after transport.
Thus the SMB and BMB applied during a given time step depend on the ice extent
and thickness at the start of the time step rather than in the middle (after transport).
Making thickness changes a function of the values at the start of the time step
is more in the spirit of Euler timestepping.

I added some tracer halo updates after glissade_add_smb and before horizontal transport,
since the remapping requires accurate fields in halo cells. This could be avoided by
doing all the halo updates before glissade_add_smb, and looping over all cells
(including halo cells) in glissade_add_smb. That change is deferred to a future commit.

Switching the order of operations is answer-changing for any problem with nonzero SMB or BMB.
I checked that answer changes for MISMIP+ are small, ~1% or less after several years.
The switch is BFB for the dome problem, which has zero SMB and BMB, and for other standard
LIVV tests (most of which are diagnostic only).