Set bmlt_float = 0 for cells that are not floating
CISM now sets bmlt_float = 0 wherever floating_mask = 0.
Thus, if we happen to compute nonzero bmlt_float for a grounded cell,
it will not be applied.

For this to work correctly, basal melting must be applied before the floating
mask changes. This is the case following a recent commit in which the SMB/BMB
calculation was moved before horizontal transport.

This commit would be answer-changing only for a run where nonzero bmlt_float
was erroneously being applied to grounded cells.