Created a grounding-line module from existing subroutines
I created a new module, glissade_grounding_line.F90.
The module consists of two subroutines: glissade_grounded_fraction,
which previously was in glissade_masks.F90, and glissade_grounding_line_flux,
which previously was in glissade.F90.

Parts of this module will be redacted in the CESM2.1 release.

This commit is not answer-changing.