Switched to new formula for lateral load at marine boundaries
Previously, I used the following formula to compute the lateral load
at a marine cliff face:

   p_av = 0.5d0*rhoi*grav*h_qp * (1.d0 - rhoi/rhoo)

This formula is valid for floating ice adjacent to the ocean, but not
for grounded ice adjacent to the ocean.

I switched to a more general formula that is valid for any edge between
an active ice-covered cell and an ocean cell:

   p_av = 0.5d0*rhoi*grav*h_qp &                                   ! p_out
        - 0.5d0*rhoo*grav*h_qp * (1.d0 - min(s_qp/h_qp,1.d0))**2   ! p_in

The first formula can be derived from the second when the ice is afloat.

This commit is answer-changing for any problem with grounded ice adjacent to the ocean.

This commit is answer-changing for many marine ice sheet simulations,
including the standard LIVV confined-shelf and circular-shelf tests.

I compared old and new LIVV shelf solutions and checked that the changes are as expected
(very small for circular-shelf; modest and local for confined-shelf).

I also ran a few time steps for MISMIP+ and verified that answer changes are small.