Cleaned up gradient_margin options, changed default from 1 to 3
CISM supports four ways to compute edge gradients at ice margins, where there
is ice on one side of the edge but not the other. With this commit,
the four methods are labeled as follows:

HO_GRADIENT_MARGIN_ALL = 0
- Compute the edge gradient when either cell is ice-covered

HO_GRADIENT_MARGIN_GROUNDED_ICE = 1
- Compute the edge gradient for grounded ice above ice-free land or ocean
 (but not for floating ice above ice-free ocean)

HO_GRADIENT_MARGIN_ICE_ONLY = 2
- Compute the edge gradient only when both cells have ice

HO_GRADIENT_MARGIN_ICE_OVER_LAND = 3
- Compute the edge gradient for ice-covered cells above ice-free land,
  but not for ice-covered cells above ice-free ocean

With this commit, option 3 becomes the default. Based on whole-ice-sheet runs
with both land- and marine-based ice, it is the most accurate and robust option.

Option 1 was the old default, but it overestimates the forcing for grounded marine-based
ice cliffs. At these cliffs, CISM's HO solver now computes a lateral load, which
provides a spreading force in the absence of a surface elevation gradient.

CISM's SIA solver, however, does not compute a lateral load. If the user tries to run
with the SIA solver and the new default gradient_margin option 3, CISM will default
to option 1 and write a log message to that effect.

I modified comments in several modules to clarify the differences between options.

This commit is BFB for standard LIVV tests, including the dome problem,
because options 1 and 3 currently give the same gradients at land-based margins.
However, a followup commit will change the way gradients are averaged from edges
to vertices for option 3, resulting in answer changes.