Changed averaging method for gradient_margin option 3
The various gradient_margin options involve computing gradients on edges
and averaging them to vertices. When the gradient is nonzero at one edge
and zero at the other, we can either
(1) take the nonzero edge gradient as the vertex value
(2) average the nonzero and zero edge gradients to get the vertex value.

With this commit, gradient_margin option 3 (the new default) switches from
method (1) to method (2). This leads to small answer changes near margins.

This commit is answer-changing for the dome problem. Changes are small, as expected.