Added timescale for limiting marine cliffs
When limit_marine_cliffs = .true., CISM compares the thickness of marine cliffs
to the theoretical maximum from Bassis & Walker (2012). Any excess results
in thinning; the thinned ice is added to the calving field.

Previously, the thinning time scale for cliffs was set to calving_timescale.
However, we may want different timescales for cliff thinning as opposed to calving.
This commit adds a cliff_timescale variable to the calving derived type.
The new variable is passed to the main calving subroutine.

The default is cliff_timescale = 0, which means that the ice thickness
is immediately set to the theoretical maximum during each timestep.
This setting appears to be stable. But we may want to try different values
when we look at marine ice cliff instability.

I ran some 4-km Greenland tests with limit_marine_cliffs = .true. and found that
this setting reduces some large velocities (~8 km/yr) at a few vertices where a
marine-based cliff is adjacent to open water. This might allow a longer time step.

This commit is answer-changing only for runs with limit_marine_cliffs = .true.
Standard tests are BFB.