Always allocate bmlt_applied and calving damage arrays
Previously, the calving damage and damage_column arrays were allocated
only when using the CALVING_DAMAGE option. However, some compilers
(e.g., nag) require them to be allocated before being passed to
the calving subroutine.

With this commit, the two arrays are always allocated.
When not using the CALVING_DAMAGE option, they are allocated
with size 1 in each dimension so as not to waste memory.

Also, bmlt_applied is now always allocated.
It is computed only for the glissade dycore, but since it appears in
glide_diagnostics.F90, it needs to be allocated for either glide or glissade.

In some future cleanup, it might be better to have separate diagnostic
subroutines for glide and glissade.