Compute global mass budget diagnostics for Glissade only
For Glide runs, I was getting a seg fault because the diagnostic subroutine
was trying to use an array that had not been allocated.
Rather than allocate the array, I added logic to compute global mass diagnostics
(i.e., global sums of SMB, BMB, calving, etc.) for Glissade but not for Glide.

Some of the required fields are not currently computed for Glide, and
it doesn't seem important at this point to add this capability to Glide.