Added an inversion scheme to estimate traction parameters and bmlt_float
This is the first version of a scheme to invert for basal traction parameters
and sub-shelf melting by relaxing the ice toward a target ice thickness.
The method is similar to that of Pollard & DeConto (2012).

I added a new config option, which_ho_inversion, with three values:
(0) HO_INVERSION_NONE. Default behavior; code works as before.
(1) HO_INVERSION_COMPUTE. Compute basal traction parameters C_p and C_c,
    aka powerlaw_c and coulomb_c, for each grid cell so as to relax the
    ice thickness (thck) toward the observed thickness (thck_obs) read from
    the input file. At the same time, compute a bmlt_float field that
    restores the thickness of floating cells to thck_obs each time step.
(2) HO_INVERSION_PRESCRIBE. For grounded grid cells, use C_p and C_c as
    computed from a previous inversion and read from the input file.
    For floating cells, use bmlt_float from a previous inversion.

Option (1), but not (2), is implemented with this commit. It works for the
Schoof basal sliding law (which_ho_babc = 11), but not for other sliding laws.
It could be implemented later for the Tsai sliding law.

To implement this option, I added several 2D fields which can be included in I/O:
- thck_obs = observational target thickness field
- powerlaw_c_2d = 2D field of C_p
- coulomb_c_2d = 2D field of C_c

Recall that for the Schoof law, C_p dominates (giving power-law behavior) in most regions,
but C_C can dominate (giving Coulomb behavior) where N is small and u is large
(e.g., near the grounding line).

The target thickness, thck_obs, can be read from the input file. If this field
is not present in the input file, CISM sets thck_obs = thck at initialization.

The powerlaw_c_2d field can also be read from the input file.
If not present, CISM sets it to a sensible default value, basal_physics%powerlaw_c
(a constant already used for the Schoof sliding law).

During the run, powerlaw_c_2d is prognosed, and then couloumb_c_2d is diagnosed as
powerlaw_c_2d/powerlaw_coulomb_ratio, where the ratio is a prescribed parameter.
Then coulomb_c_2d is limited to be <= 1, so the basal shear stress will not exceed the
effective pressure N.

I added a new subroutine, calc_basal_inversion, to prognose C_p during the run.
Evolution of C_p is described by

          dC_p/dt = (C_p/t_inv) * [-(H - H_obs)/H_inv - dH_dt/dH_dt_inv].

Here, t_inv is a time scale for inversion
      H_inv is a thickness scale (set to 50 m for now)
      dH_dt_inv is a growth rate scale

If the ice is thicker than the target, and/or the thickness is increasing,
C_p decreases to make the bed more slippery and reduce the thickness.
If the ice is thinner than the target, and/or the thickness is decreasing,
C_p increases to make the bed rougher and increase the thickness.
The two terms in brackets can compete with each other, with their relative strength
controlled by the parameters H_inv and dH_dt_inv.

After C_p is computed, it is smoothed with a Laplacian smoother that
suppresses B-grid noise and speeds convergence.
The smoother has a factor that determines how strong the smoothing is.
For relatively strong 4-1-1-1-1 smoothing, this factor would be 0.125.
With smaller values, the values in neighboring edge cells have smaller weights.

I set the default value to 0.05, which usually is large enough to suppress noise.
Once the inversion has run for a while and the C_p field is sufficiently smooth,
it may be desirable to continue the inversion with a smaller smoothing factor.
With a large factor, spatial gradients in C_p cannot be as large. As a result,
it may not be possible to adjust C_p to its optimal value.

In some cases the smoothing may be strong enough to reverse the sign of the change in C_p.
For instance, the thickness error may call for reducing C_p, but if the cell lies in
a "bowl" of low C_p, the smoother may push C_p in the opposite direction.
In this case, C_p is held fixed. That is, the smoother can halt the decline,
but it is not allowed to drive C_p in the other direction.

It is assumed that C_p and C_c are independent of basal temperature and basal water depth.
Any dependence of basal stress on basal temperature or water depth would be included in N,
or implicitly in the basal velocity.

There are several new config parameters for inversion. Parameters and default values are:
* t_inv = 200 yr
* H_inv = 50 m
* dH_dt_inv = 0.5 m/yr
* powerlaw_c_min = 2.e3 Pa (m/yr)^(-1/3)
* powerlaw_c_max = 1.e5 Pa (m/yr)^(-1/3)
* powerlaw_coulomb_ratio = 2.e4 Pa (m/yr)^(-1/3)
* smoothing factor = 0.05

These values give good results for a MISMIP+ test case but may not be optimal
for more realistic problems.

The scheme is intended for application to marine ice sheets, not just land-based ice.
In order to get good behavior near the GL, I found that where the observed ice is floating,
I needed to prescribe basal melting to maintain thck = thck_obs.
Where the observed ice is grounded, the thickness is free to evolve.
For some experiments, initially grounded ice becomes floating, but it later re-grounds,
provided the run is reasonably well behaved.

I applied the inversion scheme to a MISMIP+ problem spun up for 20 ka using the
Schoof sliding law, with N computed as in Leguy et al. (2014) using p = 1.
The spinup has C_p = 10000 Pa (m/yr)^(-1/3) and C_c = 0.50 everywhere.
I initialized C_p to values ranging from 5000 to 20000.
I verified that over several millennia, C_p and C_c gradually converge to the MISMIP+ values,
with thck converging to thck_obs to a very good approximation.
After ~2 ka, thck - thck_obs << 1 m everywhere, and C_p is within << 1% of the target value.

Minor changes:

- I changed HO_BABC_COULOMB_CONST_BASAL_FLWA to HO_BABC_COULOMB_POWERLAW_SCHOOF.
  For a while, we've been referring to this sliding law as the Schoof law.

- I changed 'coulomb_C' to 'coulomb_c' and 'powerlaw_C' to 'powerlaw_c' wherever they appear.
  Bill Sacks pointed out that config parameters with upper-case characters can be mishandled.

Answers change only for configurations using the new inversion option.

Future work:
(1) Support inversion for the Tsai sliding law in addition to Schoof.
(2) Implement and test which_ho_inversion = 2: Prescribe 2D fields of C_p and C_c
    based on values obtained from inversion.
(3) Apply to problems with prognostic temperature. (MISMIP+ has temperature and flwa prescribed.)
(4) Apply to more realistic geometry (e.g., all Antarctica).
(5) Optimize the inversion parameters.