Update builds bash scripts for POSIX compliance
The builds/*/*.bash scripts, which allow out-of-source builds, have been
update to be posix compliant. The .bash extension has been changed to
.sh to reflect these changes.

A README has been added to explain the differences between the .sh build
scripts and the extension-less build scripts (which should work in any
shell, but don't allow out-of-source builds).

Likewise, the .bash extension has been dropped from the
tests/regressions/setup_* scripts which are also shell agnostic.