Added an SSA flow factor for floating ice
Since Glimmer days, the code has had a flow enhancement factor that pre-multiplies A.
Its default value is 1.0.

With this commit, there is an additional parameter, flow_enhancement_factor_ssa,
that pre-multiplies A for floating ice.  Its value is 1.0 by default, but it can be
set to a different value in the config file.

For Antarctic runs, flow_enhancement_factor_ssa will typically be set to a value less than 1.0
to reduce the speed in ice shelves, which otherwise are too fast.
For example, Pollard & DeConto (2012, GMD) use a factor E = 0.3.
I found in Antarctic runs that a value of 0.3 to 0.4 gives good agreement with
observed shelf velocities.

This commit is answer-changing only when the user sets the factor to a value different from 1.