Added a subroutine to find interior lakes
I wrote a subroutine to find interior lakes. These are defined as connected
regions where ice is present and floating, but there is no path through other
floating cells to ice-free ocean. The logic is similar to the algorithm
used to remove icebergs.

This subroutine is called at the start of the inversion procedure,
when CISM computes a mask to identify floating cells whose thickness
should be restored to observations. Lakes are now excluded from this mask,
since it is unphysical to apply ocean melting to ice that is isolated
from the ocean.

For now, the subroutine glissade_find_lakes is in the calving module,
just before the flood-fill utility subroutine. Later, it should probably
go somewhere else.

I tested the new code on idealized lake geometry, and then on the 8-km
Antarctic grid. As far as I can tell, it correctly identifies all lakes.