Switched location of initial calving mask call
For the calving mask option (whichcalving = 5), subroutine glissade_calving_mask_init
is called at initialization to identify cells where any floating ice is assumed to calve.
I moved this call from just before to just after the initial calving call.
The reason is that the initial calving can cull cells at the calving front, and
we want the calving mask to reflect the calving-front location after the culling.

This turns out to be important for Antarctic inversion, because restoring the
ice thickness in culled cells can lead to dynamics instabilities.

This commit can be answer-changing for runs that use a calving mask and also
have initial calving.