Made calving-front cells active if adjacent to land
When the subgrid calving front scheme is turned on (which_ho_calving_front = 1),
most calving-front cells are inactive. Previously, CF cells have been active only if
their thickness is greater than or equal to the thickness of upstream interior ice.

With this commit, calving-front cells are also active if they are adjacent to land cells.
This change fixes a numerical instability that turned up during an Antarctic inversion.
Recall that the purpose of the CF scheme is to avoid unstable, inaccurate dynamics
resulting from a thin CF cell adjacent to a much thicker cell (giving a large gradient
in surface elevation). When the upstream cell is land-based (and by definition grounded),
a large elevation gradient between adjacent cells is not likely to destabilize the model,
and might be realistic.

This commit is answer-changing for marine ice simulations with calving-front cells
adjacent to land, if the CF scheme is turned on.