Do not invert for bmlt_float in cells adjacent to land
With this commit, there is a second exception to the rule that we invert
for bmlt_float in cells that are initially floating: we do not invert
in cells adjacent to land. Inverting for bmlt_float in cells adjacent
to land can result in instabilities in Antarctic runs.

(The other exception is for lake cells, which have no connection to the ocean.)