Changed inversion method for bmlt_float
Previously, the bmlt_float_inversion field was computed as the net melt rate
needed to immediately restore the model thickness in initially floating cells to the
observed thickness.

With this commit there are several changes:
- Inversion is applied not to cells that were initially floating (i.e., in the observed
  data set), but to cells that are floating (and are not lakes) at the current time step.
  I verified in a MISMIP+ test case that the GL converges to the desired location
  even when initially-floating-but-now-grounded cells have bmlt_float_inversion = 0.
- Instead of restoring the thickness in all floating cells at each time step,
  the melt rate is relaxed toward the target thickness with a prescribed time scale.
- After bmlt_float-inversion is initially computed, Laplacian smoothing is applied
  with a smoothing factor. This is similar to the smoothing done for the powerlaw_c field.
  It is hoped that the smoothing will result in a smoother bmlt_float_inversion
  field for Antarctic simulations.

To support these changes, I wrote subroutine invert_bmlt_float in the inversion module.
I added user-configurable parameters inversion_bmlt_timescale and inversion_bmlt_smoothing_factor,
with default values (for now) of 10 yr and 0.01, respectively.
I inserted '_babc' in the names of some existing parameters used to invert for powerlaw_c.

I also fixed a minor bug. In previous runs, the smoothing factor for babc was supposed
to be user-configurable, but in fact the default value in glide_types was not changed.

I tested these changes in a MISMIP+ test case. In a run without inversion, I applied
a small, uniform basal melt rate to floating cells and ran to steady state.
Then in a run with inversion, I was able to recover the initial melt rate.

This commit is answer-changing only for runs with inversion.