Added and renamed some inversion fields
For which_ho_inversion = 2, the prescribed fields from a previous inversion
are now called powerlaw_c_prescribed and bmlt_float_prescribed.

For which_ho_inversion = 1, the powerlaw_c and bmlt_float fields obtained by inversion
and written to the restart file are now called powerlaw_c_inversion and bmlt_float_inversion.

With this commit, the tavg versions of these variables are also written to the restart file.
At the end of the inversion spin-up, the tavg variables can be renamed with the 'prescribed' suffix
and read in for the forward run.

The purpose of these changes is to make variable names consistent and to simplify
the transition from the inversion spin-up to the forward run.
With these changes (if I haven't left out anything), the restart file from the spin-up
can be used to start the forward run with no changes other than renaming two variables
and the file itself.

Another change: I removed effecpress from the restart file, and I removed C_space_factor
for which_ho_babc options that do not use C_space_factor. This could have been done earlier,
but I did it with this commit to avoid having an overly long list of restart variables.
At some point we may need to allow a longer restart variable list.

Note: While testing the code, I discovered that time averages are computed correctly for tavg
variables only if these variables are listed in exactly one output file. If listed in more
than one output file (including the restart file), the accumulation and averaging are incorrect
because the accumulation subroutine is called more than once per time step.
I did not see an easy way to fix this, but I added a comment in the template for glide_io.F90.