Added code to set powerlaw_c in forward runs with inverted parameters
This commit includes a new subroutine, prescribe_basal_traction, that sets powerlaw_c
in all cells for forward runs.

This subroutine is needed because the grounding line can advance and retreat in forward runs,
relative to its position at the end of the inversion run.
In this case we should set powerlaw_c to a value other than what was computed during the inversion.

Let Cp = the value to be set, and Cp_presc = the prescribed value from the inversion run.
The rules are as follows:
(1) Read in Cp_presc at initialization, and set Cp = Cp_presc.
(2) For grounded cells at runtime:
    (a) If Cp > 0, leave Cp as is.
    (b) If Cp = 0, either set Cp = Cp_presc (if Cp_presc /= 0), or else extrapolate Cp
        from a neighbor cell. In most cases a nonzero neighbor value will exist,
        because the GL is unlikely to advance more than 1 cell per time step.
(3) For ice-free and floating cells at runtime, set Cp = 0.
    If a floating cell later regrounds, follow rule (2).

The time-varying Cp is placed in array powerlaw_c_inversion.
The time-independent Cp_presc is placed in array powerlaw_c_prescribed.
For forward runs, both these fields are written to the restart file.

A related change: During the inversion run, Cp is now set to zero in floating cells,
instead of being set to a nonzero default value. When a floating cell regrounds,
Cp is set to a nonzero default value and is adjusted from there.
Thus, when Cp is written to the restart file, we have Cp = 0 for all floating cells.
If this field is subsequently read in for a forward run, floating cells have Cp_presc = 0,
instead of a default value that might be unrealistic.

I ran a multistage MISMIP+ test as follows:
(1) 2ka spinup with Cp = 1.0e4 and bmlt_float = 0.30 m/yr
(2) 2ka inversion run, attempting to recover Cp and bmlt_float given the target thickness from (1)
(3) 1 ka forward run, given inverted Cp and bmlt_float from (2)

I verified that (2) does a good job of recovering Cp and bmlt_float while maintaining the GL position.
Also, I verified that (3) keeps the GL in place and does not allow large changes in thickness.

The next step is to test these changes for an Antarctic run.