Fixes for the Halfar test case
When I rewrote some I/O logic last summer, I introduced a potential problem:
the code crashes with a divzero if the output frequency is specified to be less
than half the model timestep (because a 'nint' calculation yields a zero).
With this commit, the code checks for output frequency nfreq = 0.
If found, nfreq is set to 1 and a warning is written to the log file.

In the Halfar test directory, I changed the output frequency to 5 years
(since dt = 5 years) and modified a couple of comments.

With this fix, the Halfar test works as before for both SIA and HO solvers.