Revised and cleaned up the gradient operators
I'm making another pass through the gradient subroutines in glissade_grid_operators.F90,
cleaning things up for the release.

I introduced a temporary optional argument called new_gradient, defined in
glissade_velo_sia.F90 and glissade_velo_higher.F90. When this argument is set to false,
gradient computations are as before. When set to true, there are several changes:

- In glissade_velo_sia.F90, surface elevation gradients are computed by calling
glissade_gradient, which is a slightly simpler version of glissade_centered_gradient.
It computes a 2nd order centered gradient, but without the optional max_slope argument.
I made a similar replacement for gradient computations in the L2L2 solver.

- I wrote a new subroutine called glissade_surface_elevation_gradient_new,
to replace glissade_surface_elevation_gradient. With the new logic, this is the
only subroutine needed to handle the various options for which_ho_gradient and
which_ho_gradient_margin (and higher-order surface elevation gradients generally).
Previously, this subroutine was used only for gradient_margin option 3, and
other subroutines were used for options 0 to 2.

- This new subroutine includes code to compute upstream gradients, not just centered.
Previously, without the upstream code, the gradient was always computed as centered
for which_ho_gradients_margin = 3, even when which_ho_gradient = 1 (for upstream).
As a result, the dome test always gave checkerboard noise after ~100 years.
Thanks to Heiko Goelzer for pointing out this bug.

- I added a third option to which_ho_gradient. The options are now (0) centered,
(1) first-order upstream and (2) second-order upstream. Previously, first-order
and second-order upstream were distinguished by a hard-coded flag.
Note that for the dome test, (1) and (2) appear to give similar results.
For the Halfar test, however, (2) is significantly more accurate.

I ran several Halfar tests with the old and new gradient subroutines to make sure the
new versions did not unexpectedly change answers. For this commit, I left the old
subroutines in place, so we can go back and run more comparisons if needed,
simply by setting new_gradient = false.
The next commit will remove the old logic and subroutines.

Note: In the committed code, new_gradient = .false.
With this setting there is no change in the gradient computations, and answers are BFB
for the standard dome test and other standard LIVV tests.
With new_gradient = .true., there are significant changes (~1 m/yr) in the confined-shelf test,
and roundoff-level changes for the ISMIP-HOM tests (A, C and F) and stream test.

Answers change substantially when an upstream gradient is used with
 which_ho_gradient_margin = 3, because of the bug fix.