Add initialization with unstaggered temperature from glide
Add a new variable tempunstag, which has two functions:
- supplied in restart or input file instead of tempstag, it is used
  to initialize ice temperature by interpolation to tempstag
- diagnostic output of temperature on unstaggered grid

This addition makes it possible to use a glide temperature field as
input to a glissade run, which is interpolated from the unstaggered
vertical grid to the staggered vertical grid.

If present, the variable tempunstag is automatically read from input files,
but is only used for temperature init if tempstag is not present. If specified
as output variable, it contains the temperature interpolated from tempstag to
the layer interfaces. tempunstag and tempstag are identical at the top and
bottom boundaries.