Modified the implementation of beta for ISHOM-C test case
In glissade_basal_traction.F90, under the case "HO_BABC_ISHOMC", beta is now computed
on the global grid and the test case can be run in parallel.
I added two allocatable arrays "beta_global" and "unstagbeta" for the computation.

In glide_setup.F90 I deleted the reference to the FATAL error message that
would not permit the run of ISMIP-HOME C on more than 1 processor.

The results of a run with 4-proc are within roundoff from a run with 1 proc.