Allocate global beta array on main task only
Instead of allocating a global beta array on all tasks, beta_global is now allocated
globally on the main task only, and allocated with size (0,0) on other tasks.
This workaround is needed since distributed_scatter_var tries to deallocate
the input array on all tasks. (Maybe something to fix later.)

Also, I changed 'unstagbeta' to 'beta_extend', since the extended beta field is
co-located with beta on the velocity grid, but with ice_grid dimensions.