Make which_ho_babc = 8 the default for ISMIP-HOM C
Now that options which_ho_babc = 8 has been parallelized, it should be the default
for running ISMIP-HOM test C, since it is more accurate than reading beta from
an external file.

(This goes back to the long-standing issue that we cannot read in global arrays
on the velocity grid if they have nonzero values at the global boundary.
Reading in a global beta array on the ice grid introduces interpolation errors.
So it is better to construct beta on the velocity grid at runtime.)

NOTE: This commit is answer-changing for ISMIP-HOM C, so LIVV users should create
a new reference directory.  Other standard tests are BFB.