Fixed computation of fyear in libglint/glint_example_clim.F90
fyear used to be normalized to 1 meaning that only the first year of forcing data would be read. Now if the length of the run is longer than the available forcing time series the code loops through these forcings.
Note: Ran test to make sure nothing was broken using glint-example test case and modifying config options.