Adding cell_area variable for CMIP6 output
I added a variable called cell_area which returns a field of cell areas as requested by the CMIP6 variable output.
The variable is defined in glide_var_def.F90, its type is set in glide_type.F90 and its computation is done in glide.F90 and glissade.F90.
The variable is simply the multiplication from the width by the length of a given cell. It happens to be constant in CISM but maybe in the future this could change.