Addition of MISMIP3d as a test case
Adding MISMIP3d as a test case in the CISM test case directory.
Right now the config file default assumes that which_ho_babc = 10 which for now is the only
parameter allowing a modification of C_space_factor.

Note: this will most likely be modified in the next couples of days or after the release
in which case a slight modification will need to happen to the file mismip3dSetup.py.