Update MISMIP3d and MISMIP+ test cases.
MISMIP3d:
fixed the run script to allow clean restart for experiments P75S and P75S. The previous script
did not take into account the removal of a command line option.
The setup script has only minor and esthetic changes.

MISMIP+:
Added restart capabilities to the Run script for all experiments. This include the automatic reset
of internal_time in the Spinup restart file when running Ice0, Ice1r and Ice2r for the first time.
The reset will only be done if the run script is used.
The setup script has only esthetic changes.
The README file was modify adding guidance on how to properly use the restart capability while
using the run script.