Compute flwa for all cells, including ice-free cells.
Previously, the flow factor flwa was computed only for cells with ice (ice_mask = 1).
Now, flwa is computed for all cells, including ice-free cells.
This ensures that flwa has physically sensible values in ice-free cells
at the ice margin, in case these values are needed when interpolating flwa
to the staggered grid.

This assumes that ice-free cells have a sensible default temperature such as
the air temperature (artm), as is currently the case.