Changed f_flotation calculation for ice-free land cells
This commit changes the way f_flotation is computed and extrapolated for ice-free land,
when using the linear flotation function (which_ho_flotation_function = 2) for the GLP.

Recall that for the linear function, f_flotation is equal to the sub-shelf cavity depth
for floating ice. With this commit, f_flotation is set to -(topg - eus) for ice-free land.
(Previously it was set to a large negative value, -10000.)
This gives a smooth change in f_flotation when thin ice crosses the thklim threshold
and activates.  It also gives a more accurate interpolation to vertices that have both
ice-covered floating neighbors and ice-free land neighbors.

When f_flotation is extrapolated, the ice-free land values are now treated as physically
realistic, so only ice-free ocean cells require an extrapolated value.

This commit is answer-changing for runs with a GLP.