Require f_flotation < 0 for land cells
For land cells, the flotation function is given by -(topg - eus), where usually eus = 0.
This is problematic for cells with topg	= 0. These are defined to be land cells, but
the f_ground computation can yield NaNs	when f_flotation = 0 in	adjacent land cells.

With this commit, the value of (topg - eus) used in evaluating f_flotation is
limited to be >= 1 m. This give f_flotation < 0 for all land cells and fixes NaNs.