Include ice-free land cells in stagthck
The higher-order velocity solver uses stagthck (the ice thickness interpolated
to the staggered mesh) for many calculations. Previously, ice-free land cells
were not included in the interpolation.

With this commit, the interpolation includes ice-free land cells. That is, stagthck
is the average thickness of the neighboring cells that have ice and/or are land-based.
Ice-free ocean cells are still excluded from the average.

This commit is answer-changing for the dome test and for most other simulations
with a land margin. The changes are modest: ~1 m/yr in velocity near the dome margin,
with a cumulative thickness difference of ~1 m after 10 years.
Users should create a new reference directory after this commit.

For now, the new staggering can be turned off by setting new_stagger = F
at the top of glissade_velo_higher. I checked that answers in this case are BFB.