Changed local assembly of taud and beta terms
Glissade has two options for assembling the taud and beta terms: standard and local.
(Recall that the taud term includes surface gradients and goes on the RHS,
and the beta term includes basal traction and goes into the matrix.)
This commit modifies local assembly.

The intent of local assembly is to make the solution at a given vertex dependent
on the surface gradient and beta values at that vertex, and no other vertices.
In standard FE assembly, values from other vertices arise from loops over elements
(which include neighbor vertices). With local assembly, neighbor vertices are excluded.

Previously, the local terms depended on the number of active cells surrounding
the vertex, and could change abruptly when a cell transitioned from thck < thklim
to thck > thklim. With this commit, the local terms are computed in the same way
regardless of whether adjacent cells are active or inactive.
This gives smoother changes in the local terms and can prevent dithering at the margin
(where a cell oscillates between active and inactive as a result of oscillations
in the values of local terms). The local terms are now computed in loops over vertices
rather than active cells, and thus the computation is simpler.

For now, the new local assembly can be turned off by setting new_assembly = F
at the top of glissade_velo_higher.  With new_stagger and new_assembly = F,
standard LIVV tests are BFB compared to two commits ago (before the stagger
and assembly changes).

This commit (with new_assembly = T) is answer-changing at roundoff level
for the ISMIP-HOM and stream tests, which specify local assembly for beta.
(Standard assembly is still the default setting for other tests.)
The roundoff-level changes result from changes in the order of operations;
for these tests there are no significant changes in matrix terms.
The LIVV reference directory should again be updated.