Modification to MISMIP and MISMIP+ scripts
I made some minor modification to the MISMIP and MISMIP+ scripts.
MISMIP:
In mismipSetup.py there was a typo in linking the input file for the first retreat experiment.
mismipPlot.py now allows the plot of the grounding line location for unfinished experiments.

MISMIP+:
The README file reflects the changes made to other files.
mismip+Setup.py allows the setup of Ice1, Ice2 or Ice0+Ice1+Ice2 (allIce) experiments separately from the Spinup and Ice0.
This is useful when a user wants to setup a full set of experiments using different time step as in the Spinup.
mismip+Run.py needed one conversion from a string to a float.
mismip+Plot.py now automatically scales the figures based on the min and max of its entries.