Simplified the calving interfaces
The interface for subroutine glissade_calve_ice now includes the object model%calving,
which reduces the number of arguments that need to be passed in and out.

Also, I removed the direct call to glissade_calve_ice at initialization.
Instead, there is a call to glissade_calving_solve (with an init_calving flag
to handle calculations that differ at initialization), which calls glissade_calve_ice.

Some variables in the calving derived type have been renamed to avoid
redundant 'calving' in the name.

This commit is BFB.
A new configure is required because of changes in glide_vars.def.